requirements.txt
instagrapi
python-dotenv
pillow